+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- keycloak_setup: Now creates an `oauth2-proxy` Client and associated Secret
+  with the client secret for use by OAuth2-Proxy.
 
 ## [0.12.6]
 - keycloak_setup: Initial release for CSM 1.0